Refactor easycert command to a subcommand for cert generation
